fix: Pass doubles directly to pdfx render() method
- Remove .toInt() conversion - render() expects double, not int
- page.width and page.height are already numeric, multiply by 2.0 directly

+++ lib/core/services/pdf_to_png_converter.dart
@@ -28,9 +28,10 @@ class PdfToPngConverter {
 
       // Render page to image at high DPI (approximately 288 DPI)
       // Use 2.0 scale factor for good quality
+      // Pass doubles directly to render() - it expects double, not int
       final image = await page.render(
-        width: (page.width * 2.0).toInt(),
-        height: (page.height * 2.0).toInt(),
+        width: page.width * 2.0,
+        height: page.height * 2.0,
         format: PdfPageImageFormat.png,
       );
